+++ app/services/ai/package.json
@@ -18,6 +18,7 @@
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
+    "node-cron": "^3.0.0",
     "node-fetch": "^2.6.1"
   }
 }
