+++ frontend/src/styles/Ai/Ai.scss
@@ -5,8 +5,9 @@
     margin: 4vh 0vw 0vh 4vw;
 }
 
-.header {
+.ai-header {
     text-align: center;
+    font-size: 4vw;
 }
 
 .mean{
